fix(example-app): reorganize actions
Some actions need to be defined within the scope of changes tree view, not to conflict with other actions, like Delete in the editor.
