+++ packages/example-app/src/VersionControl/Changes/useChangesViewActionDefinitions.tsx
@@ -1,10 +1,13 @@
 import React from "react";
 import { useRecoilCallback, useRecoilValue } from "recoil";
-import { ActionDefinition, PlatformIcon } from "@intellij-platform/core";
+import {
+  ActionDefinition,
+  CommonActionId,
+  PlatformIcon,
+} from "@intellij-platform/core";
 
 import { notImplemented } from "../../Project/notImplemented";
 import {
-  changeListsUnderSelection,
   changesTreeNodesState,
   changesUnderSelectedKeys,
   ChangesViewTreeNode,
@@ -14,7 +17,7 @@ import {
   queueCheckInCallback,
   selectedKeysState,
 } from "./ChangesView/ChangesView.state";
-import { allChangesState, useSetActiveChangeList } from "./change-lists.state";
+import { allChangesState } from "./change-lists.state";
 import path from "path";
 import { IntlMessageFormat } from "intl-messageformat";
 import { getExpandedToNodesKeys } from "@intellij-platform/core/utils/tree-utils";
@@ -55,7 +58,8 @@ export const useChangesViewActionDefinitions = (): ActionDefinition[] => {
     },
     {
       id: VcsActionIds.REFRESH,
-      title: "Refresh",
+      title: "Refresh VCS Changes",
+      description: "Refresh VCS Changes",
       icon: <PlatformIcon icon="actions/refresh.svg" />,
       actionPerformed: refresh,
     },
@@ -63,12 +67,6 @@ export const useChangesViewActionDefinitions = (): ActionDefinition[] => {
       id: VcsActionIds.SHOW_DIFF,
       title: "Show Diff",
       icon: <PlatformIcon icon="actions/diff.svg" />,
-      actionPerformed: notImplemented,
-    },
-    {
-      id: VcsActionIds.SHELVE_SILENTLY,
-      title: "Shelve Silently",
-      icon: <PlatformIcon icon="vcs/shelveSilent.svg" />,
       actionPerformed: notImplemented,
     },
     {
@@ -200,20 +198,8 @@ function useJumpToSourceAction(): ActionDefinition {
   };
 }
 
-const deleteChangeListMsg = new IntlMessageFormat(
-  `Delete {count, plural,
-    =1 {Changelist}
-    other {Changelists}
-  }`,
-  "en-US"
-);
 function useChangeListActionDefinitions(): ActionDefinition[] {
-  const selectedChangeLists = useRecoilValue(changeListsUnderSelection);
   const selectedChanges = useRecoilValue(changesUnderSelectedKeys);
-  const setActiveChangeList = useSetActiveChangeList();
-  const nonActiveSelectedChangeLists = [...selectedChangeLists].filter(
-    (list) => !list.active
-  );
 
   const actions: ActionDefinition[] = [
     {
@@ -224,40 +210,8 @@ function useChangeListActionDefinitions(): ActionDefinition[] {
       actionPerformed: notImplemented,
     },
   ];
-
-  // Actions are conditionally added instead of being disabled.
-  if (selectedChangeLists.size === 1) {
-    actions.push({
-      id: VcsActionIds.RENAME_CHANGELIST,
-      title: "Edit Changelist...",
-      description: "Edit name and description of selected changelist",
-      icon: <PlatformIcon icon="actions/edit.svg" />,
-      actionPerformed: notImplemented,
-    });
-  }
-  if (nonActiveSelectedChangeLists.length > 0) {
-    actions.push({
-      id: VcsActionIds.REMOVE_CHANGELIST,
-      title: deleteChangeListMsg.format({
-        count: nonActiveSelectedChangeLists.length,
-      }) as string,
-      description: "Remove changelist and move all changes to another",
-      icon: <PlatformIcon icon="general/remove.svg" />,
-      actionPerformed: notImplemented,
-    });
-  }
-  if (nonActiveSelectedChangeLists.length === 1) {
-    actions.push({
-      title: "Set Active Changelist",
-      description: "Set changelist to which new changes are placed by default",
-      id: VcsActionIds.SET_DEFAULT_CHANGELIST,
-      icon: <PlatformIcon icon="actions/selectall.svg" />,
-      actionPerformed: () => {
-        setActiveChangeList(nonActiveSelectedChangeLists[0].id);
-      },
-    });
-  }
   if (selectedChanges.size > 0) {
+    // FIXME: should use active paths, not selection
     actions.push({
       title: "Move to Another Changelist...",
       description: "Move selected changes to another changelist",
